Add files via upload
The quizapp application will contain questions followed by a total score shown at the end of the quiz. The score will increase based on the correct answers given. Initially there are only eight questions but you can increase questions in the JavaScript file.
We will create the structure using HTML and design it with CSS. JavaScript will be used to implement the logic and scoring of the quiz.